System.Configuration: Pull in ConfigurationElementCollection from referencesource.
